+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/model/DatabaseMappingImpl.java
@@ -48,6 +48,8 @@ import org.apache.hadoop.hive.metastore.api.LockRequest;
 import org.apache.hadoop.hive.metastore.api.MetaException;
 import org.apache.hadoop.hive.metastore.api.Partition;
 import org.apache.hadoop.hive.metastore.api.PartitionSpec;
+import org.apache.hadoop.hive.metastore.api.PartitionValuesRequest;
+import org.apache.hadoop.hive.metastore.api.PartitionValuesResponse;
 import org.apache.hadoop.hive.metastore.api.PartitionsByExprRequest;
 import org.apache.hadoop.hive.metastore.api.PartitionsByExprResult;
 import org.apache.hadoop.hive.metastore.api.PartitionsStatsRequest;
@@ -481,6 +483,12 @@ public class DatabaseMappingImpl implements DatabaseMapping {
       transformOutboundTable(table);
     }
     return result;
+  }
+
+  @Override
+  public PartitionValuesRequest transformInboundPartitionValuesRequest(PartitionValuesRequest req) {
+    req.setDbName(transformInboundDatabaseName(req.getDbName()));
+    return req;
   }
 
   @Override